Feat/#1474 validating behaviour (#1886)
* seperate types, styled component and constants from main Button file

* update Button imports

* add custom loader and pass through loading styles for Button

* pass through styles

* add custom Loader prop and change loading state component logic

* seperate Button types and styling into their own files

* add instruction box loading styling

* add copyrights
