+++ uikit/Button/constants.tsx
@@ -17,19 +17,22 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import styled from '@emotion/styled-base';
+import { ButtonVariant, ButtonSize } from './types';
 
-const FocusWrapper = styled('button')`
-  border: none;
-  background: none;
-  padding: 0px;
-  cursor: pointer;
-  box-shadow: none;
-  outline: none;
-  transition: box-shadow 0.1s ease-in;
-  &:focus {
-    box-shadow: 0px 0px 4px 0px ${({ theme }) => theme.colors.secondary_1};
-  }
-`;
+export const BUTTON_VARIANTS: {
+  PRIMARY: ButtonVariant;
+  SECONDARY: ButtonVariant;
+  TEXT: ButtonVariant;
+} = Object.freeze({
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  TEXT: 'text',
+});
 
-export default FocusWrapper;
+export const BUTTON_SIZES: {
+  SM: ButtonSize;
+  MD: ButtonSize;
+} = Object.freeze({
+  SM: 'sm',
+  MD: 'md',
+});
